requirements.txt
Flask-SQLAlchemy
flask-wtf
flask-bcrypt
Flask-Login
